Установка слушателя событий в Table_1

+++ src/main/java/ru/RawMaterial.java
@@ -9,8 +9,8 @@ import java.util.Map;
 public class RawMaterial {
 
     private String name; //краткая абривиатура сырьевого материала
-    private int priceBK5; //цена за тонну в юанях
-    private int priceBK6; //цена BK6
+    private double priceBK5; //цена за тонну в юанях
+    private double priceBK6; //цена BK6
     private double bd; //истинная плотность
 
 
@@ -28,10 +28,10 @@ public class RawMaterial {
         this.name = name;
     }
 
-    public void setPriceBK5(int price) {
+    public void setPriceBK5(double price) {
         this.priceBK5 = price;
     }
-    public void setPriceBK6(int price) {
+    public void setPriceBK6(double price) {
         this.priceBK6 = price;
     }
 
@@ -47,10 +47,10 @@ public class RawMaterial {
         return name;
     }
 
-    public int getPriceBK5() {
+    public double getPriceBK5() {
         return priceBK5;
     }
-    public int getPriceBK6() {
+    public double getPriceBK6() {
         return priceBK6;
     }
